+++ RealmSwift-swift2.0/Tests/ObjectCreationTests.swift
@@ -446,6 +446,9 @@ class ObjectCreationTests: TestCase {
     private func verifySwiftOptionalObjectWithDictionaryLiteral(object: SwiftOptionalDefaultValuesObject, dictionary: [String:AnyObject], boolObjectValue: Bool?) {
         XCTAssertEqual(object.optBoolCol.value, (dictionary["optBoolCol"] as! Bool?))
         XCTAssertEqual(object.optIntCol.value, (dictionary["optIntCol"] as! Int?))
+        XCTAssertEqual(object.optInt16Col.value, (dictionary["optInt16Col"] as! Int16?))
+        XCTAssertEqual(object.optInt32Col.value, (dictionary["optInt32Col"] as! Int32?))
+        XCTAssertEqual(object.optInt64Col.value, (dictionary["optInt64Col"] as! Int64?))
         XCTAssertEqual(object.optFloatCol.value, (dictionary["optFloatCol"] as! Float?))
         XCTAssertEqual(object.optDoubleCol.value, (dictionary["optDoubleCol"] as! Double?))
         XCTAssertEqual(object.optStringCol, (dictionary["optStringCol"] as! String?))